ai.kanggyeonggu.store/mlservice/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# 데이터 분석 및 머신러닝 라이브러리
pandas>=2.1.0
numpy>=1.24.0
scikit-learn>=1.3.0
datasets>=2.14.0
openpyxl>=3.0.0
xlrd>=2.0.0
requests>=2.31.0

#환경 변수 관리
python-dotenv>=1.0.0
folium>=0.14.0

# 시각화 라이브러리
matplotlib>=3.7.0
seaborn>=0.12.0
folium>=0.14.0

# 디버깅 및 개발 도구
icecream>=2.1.3

# 자연어 처리 라이브러리
nltk>=3.9.0
wordcloud>=1.9.0
konlpy>=0.6.0
sentencepiece>=0.2.0
jpype1>=1.6.0
